Broken: Works only with numeric usernames 1-65535 inclusive
Please, this needs to be fixed. Ugh.

Probably need to set up routine pulling of the caddy and http.git repos,
patch them with the changes that fix the URL parsing, and build them
together for a custom caddy binary.

Should also upstream the patch.

This fix is intended to last indefinitely, if I never feel like doing
the above.