Begin moving to new method of hosting website
MiaB has an undocumented and unsupported feature that lets the
configuration of the NGiNX webserver be modified:
https://discourse.mailinabox.email/t/nginx-multiple-modified-configs/2706/2

This would remove the need for modify_nginx.service and such.

Caddy Webserver v2 has been released, and has changed a bit. A community
plugin does exist for implementing the webhook functionality provided by
a plugin that this repository relied on for being able to update the
website after content was pushed to a separate repository:
https://github.com/vrongmeal/caddygit

Installing it is a little more involved.

Since the nginx configuration can be modified, and nginx can do serve
the content, the only functionality of caddy would be:

- Responding to the webhook
- Updating the repository
- Running any other commands

These are functions that could be easily covered by the very simple
project:
https://github.com/bloomberg/python-github-webhook

Currently, an uninstall playbook is being written that will remove and
undo all of the changes this playbook has made to the server.

After that, a python package will be added that uses the webhook
framework mentioned above so that, instead of installing caddy, the
following would be done:

- Install pipx
- Install this repo's custom python webhook server package
- Setup systemd for the webhook server

This isn't that much less complex than curating a custom caddy webserver
build, but does have the advantage of using components I understand.

Additionally, it doesn't require setting up a computer for compiling Go.